Fix. 역할 책임 분리 후에 각각의 객체 생성 및 수정
