--- src/domain/Lotto.js
@@ -1,22 +1,0 @@
-import { ERROR_MESSAGE_LOTTO_LENGTH } from '../constants';
-
-class Lotto {
-  #numbers;
-  #result;
-
-  constructor(numbers) {
-    this.validLottoLength(numbers);
-    this.#numbers = numbers;
-    this.#result = 0;
-  }
-
-  get numbers() {
-    return this.#numbers;
-  }
-
-  validLottoLength(numbers) {
-    if (numbers.length !== 6) throw new Error(ERROR_MESSAGE_LOTTO_LENGTH);
-  }
-}
-
-export default Lotto;
